test(webview): merge InstructionPanel suites into one file
InstructionPanelDesktopComposer.vitest.ts and
InstructionPanelLauncher.vitest.ts rendered the same component through
the same mount harness with near-duplicate helpers. Merge them into
InstructionPanel.vitest.ts with one file-scope DOM hook (the launcher's
superset module-load callback — two per-describe hooks would rebuild the
JSDOM customElements registry without re-running the cached module
registrations), one recordEvents/changeValue pair, and composer-only
helpers scoped to its describe. All test cases from both files survive.
